Fix bugs when trying to delete/get properties from session
In the first case, the value we were trying to del from the
session could be absent if we had just connected while already
connected with another account (in this case the session is
flushed mid-process).

In the second case, not sure what could have happened! But easy
to fix that way without problematic consequences.