feat: add organizer_group field to event theme type
The new field allows specifying a published SupportGroup instance that will be
automatically added as organizer when creating the event from a request of a
given theme type.
A new event action (and Event model method) allows to add a group as organizer,
automatically creating one OrganizerConfig for each group referent (optionally
limiting to only one person, as when accepting a group co-organization
invitation).